work: 엔진 설정 파일 로드, 파싱
- 엔진 설정 파일 로드, 파싱
- 파일에서 로드한 타겟 프레임 설정
- 디버깅용 FPS 출력

+++ ConsoleEngine/Game/Demo/DemoLevel.cpp
@@ -31,7 +31,7 @@ DemoLevel::DemoLevel()
 	while (!feof(file))
 	{
 		fgets(buffer, 256, file); // 한 줄씩 읽기
-		int lineLength = strlen(buffer); // 한 줄 문자열 길이값 구하기
+		int lineLength = (int)strlen(buffer); // 한 줄 문자열 길이값 구하기
 
 		// 파싱
 		for (int i = 0; i < lineLength; ++i)
